+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # QuickFFSync - Changelog
 
+## [1.0.9] - 2025-10-18
+- Screen recording now respects the "Preset" parameter value;
+- The clipboard is now preserved even after closing the program (for example, a copied ffmpeg command).
+
 ## [1.0.8] - 2025-10-15
 - Added "Batch Convert";
 - Added "Screen Record";